ในไฟล์ C 1_5 เสร็จ

+++ 6320500620_1_3.py
@@ -1,4 +1,4 @@
-N = float(input())
+N = int(input())
 re = N%3
 re2 = N%11
 print(int(re),int(re2))
